Non-cpu2cpu memory references are now handled through bus

+++ device.h
@@ -7,10 +7,10 @@
 class Device
 {
 public:
-    virtual void set8(Bus::AddressRegion* region, uint16_t address, uint8_t data) = 0;
-    virtual uint8_t get8(Bus::AddressRegion* region, uint16_t address) = 0;
-    virtual void set16(Bus::AddressRegion* region, uint16_t address, uint16_t data) = 0;
-    virtual uint16_t get16(Bus::AddressRegion* region, uint16_t address) = 0;
+    virtual void set8(Bus::AddressRegion* region, unsigned address, uint8_t data) = 0;
+    virtual uint8_t get8(Bus::AddressRegion* region, unsigned address) = 0;
+    virtual void set16(Bus::AddressRegion* region, unsigned address, uint16_t data) = 0;
+    virtual uint16_t get16(Bus::AddressRegion* region, unsigned address) = 0;
 };
 
 #endif // DEVICE_H
